directory rearrangement added missing services, interfaces and test classes code cleanup

+++ KassensystemCPP/src/domain/model/Person.cpp
@@ -1,7 +1,6 @@
-#include "Person.h"
-#include <string>
+#include "Entities.h"
 
-Person::Person(std::string name, int id)
+Person::Person(std::string name, int64_t id)
 {
 	this->name = name;
 	this->id = id;
@@ -12,7 +11,7 @@ std::string Person::getName() const
 	return name;
 }
 
-int Person::getID() const
+int64_t Person::getID() const
 {
 	return id;
 }
